TDViewerPanel,m TDImageUtilites  - updated RescaleImage() methods to force conversion of window and level to double before combining them

+++ Library/GuiComponents/TDViewerPanel.m
@@ -1375,8 +1375,8 @@ classdef TDViewerPanel < handle
         end
         
         function rescaled_image = RescaleImage(image, window, level)
-            min_value = double(level - window/2);
-            max_value = double(level + window/2);
+            min_value = double(level) - double(window)/2;
+            max_value = double(level) + double(window)/2;
             scale_factor = 255/max(1, (max_value - min_value));
             rescaled_image = uint8(min(((image - min_value)*scale_factor), 255));
         end
